Did audio stuff instead
How to use

1) Create Audio Groups
Under the AudioManager, add AudioGroups to the respective sfx/dialogue/bgm fields. The groupName is the identifier of an sfx. An AudioGroup can contain 1 or more AudioClips. For example, when the "Shoot" SFX has multiple variants, put them all under a single Audio Group and call the group something like "Shoot".

2) Playing Audio
In the AudioManager there are the functions to play the sfx/dialogue/bgm as needed (string: group name required to lookup). For SFX, use PlaySFXOneShot(string sfxGroupName);. The groupName is what you set in part 1. For Dialogue, use PlayDialogue(string dialogueId);. dialogueId here refers to the group name of the AudioGroup under the dialogue field. For BGM, same as dialogue: PlayBGM(string bgmGroupId).
It is recommended to have 1 or more sfxs per AudioGroup for sfx.
It is recommended to have only 1 dialogue per dialogue AudioGroup, and 1 or more bgm per bgm AudioGroup.

The system will automatically randomize as needed for SFX and BGM.
For SFX, previous SFX will not get cut immediately, but will continue playing once started, so recommended to keep sfx duration small.
For dialogue, once function is called, the next dialogue will interrupt the current dialogue.
For BGM, once function is called, the next bgm will interrupt the current bgm.